fix: Align all extrinsic calls with Subtensor expected formats
BREAKING CHANGES:
- set_weights/reveal_weights: Changed weights format from u64 to u16 (scale u16::MAX)
- normalize_weights: Returns Vec<u16> instead of Vec<u64>
- add_stake/unstake: Added required netuid parameter
- serve_axon/serve_axon_tls: Added required version parameter and corrected IP encoding to u128
- increase_take/decrease_take: Added required hotkey parameter
- move_stake: Added origin_netuid and destination_netuid parameters
- transfer_stake: Corrected signature to match Subtensor (destination_coldkey, hotkey, netuids, amount)
- swap_stake: Corrected signature to match Subtensor (hotkey, netuids, amount)
- mechanism functions: Changed mechanism_id from u64 to u8 (MechId type)
- reveal_mechanism_weights: Changed weights from u64 to u16, salt from u64 to u16

Fixes:
- All weight normalization now uses u16::MAX (65535) scale instead of u32::MAX
- IP addresses correctly encoded as u128 (IPv4 as u32 in u128, IPv6 direct)
- All UIDs converted to u16 where Subtensor expects Vec<u16>
- All function signatures now match Subtensor pallet expectations

Documentation:
- Added development warning to README.md
- Updated function documentation with Subtensor expected format comments

Tests:
- Updated unit_tests.rs to include root_stake field in NeuronInfo

+++ src/utils/crypto.rs
@@ -3,8 +3,9 @@ use hex;
 use sp_core::blake2_256;
 
 /// Generate a commit hash for weights using Blake2b
-pub fn commit_weights_hash(uids: &[u64], weights: &[u64], salt: &[u8]) -> Vec<u8> {
-    // Serialize UIDs, weights, and salt
+/// Uses u16 format for weights to match Subtensor's expected format
+pub fn commit_weights_hash(uids: &[u64], weights: &[u16], salt: &[u8]) -> Vec<u8> {
+    // Serialize UIDs (as u64), weights (as u16), and salt
     let mut data = Vec::new();
     for uid in uids {
         data.extend_from_slice(&uid.to_le_bytes());
